+++ vsc-excel-vba/d.vb/dictionary.d.vb
@@ -6,10 +6,10 @@ Public Class Dictionary : Implements IEnumerable
     Public Sub Add(key As String, item As Object)
     End Sub
 
-    Public Sub Add(key As Integer, item As Object)
+    Public Sub Add(key As Long, item As Object)
     End Sub
 
-    Public Function Item(index As Integer) As Object
+    Public Function Item(index As Long) As Object
     End Function
 
     Public Function Item(key As String) As Object
@@ -21,19 +21,19 @@ Public Class Dictionary : Implements IEnumerable
     Public Function Keys() As String()
     End Function
 
-    Public Function Keys() As Integer()
+    Public Function Keys() As Long()
     End Function
 
     Public Function Exists(key As String) As Boolean
     End Function
 
-    Public Function Exists(key As Integer) As Boolean
+    Public Function Exists(key As Long) As Boolean
     End Function
 
     Public Sub Remove(key As String)
     End Sub
     
-    Public Sub Remove(key As Integer)
+    Public Sub Remove(key As Long)
     End Sub
 
     Public Sub RemoveAll()
